make requested reasonable changes
